+++ lib/utils/native/native_web.dart
@@ -1,0 +1,1 @@
+export 'native_web_noop.dart' if (dart.library.html) 'native_web_impl.dart';