+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
 * initial release.
+
+## 0.0.2
+
+* added:
+  - `CommonCard` a generic Container/MaterialButton with all properties to create custom cards, with `CommonCardTheme` support (not opinionated).
+  - `CustomShadow` to create Shadows (inner-shadows and drop-shadows) with blendMode support.
+  - `RectLayoutChangedBuilder` to detect size and position of Widgets in the screen.
+  - wrapper for `usePathUrlStrategy` plugin for io and web.